examples/art-mcp/pyproject.toml
[project]
name = "art-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.14",
    "asyncio>=3.4.3",
    "mcp>=1.11.0",
    "openai>=1.74.0",
    "openpipe-art[skypilot]>=0.4.4",
    "python-dotenv>=1.1.1",
    "weave>=0.51.56",
]
